+++ __tests__/contracts.js
@@ -4,9 +4,9 @@
 
 // @flow
 
-import {Span} from 'opentracing';
-import {removeProps, TONAddressStringVariant} from '../src/modules/TONContractsModule';
-import {TONOutputEncoding} from '../src/modules/TONCryptoModule';
+import { Span } from 'opentracing';
+import { removeProps, TONAddressStringVariant } from '../src/modules/TONContractsModule';
+import { TONOutputEncoding } from '../src/modules/TONCryptoModule';
 import {
     TONClient,
     TONClientError,
@@ -16,9 +16,9 @@ import {
 } from '../src/TONClient';
 
 
-import type {TONContractLoadResult} from '../types';
-import {version} from '../package.json';
-import {ABIVersions, nodeSe, tests} from './_/init-tests';
+import type { TONContractLoadResult } from '../types';
+import { version } from '../package.json';
+import { ABIVersions, nodeSe, tests } from './_/init-tests';
 
 async function loadPackages() {
     return {
@@ -29,7 +29,7 @@ async function loadPackages() {
         SetCodePackage: await tests.loadPackage('Setcode'),
         SetCode2Package: await tests.loadPackage('Setcode2'),
         EventsPackage: await tests.loadPackage('Events'),
-    }
+    };
 }
 
 beforeAll(tests.init);
@@ -358,8 +358,7 @@ test('External Signing on ABI v2', async () => {
     });
     const runMessage = await contracts.createRunMessage(messageParams);
 
-    expect(signedRunMessage.message.messageBodyBase64)
-        .toEqual(runMessage.message.messageBodyBase64);
+    expect(signedRunMessage).toEqual(runMessage);
 });
 
 test('Should change InitState of contract', async () => {
@@ -986,10 +985,10 @@ test('Test expire retries', async () => {
     });
 
     const client = await TONClient.create({
-            ...tests.config,
-            messageExpirationTimeout: 5000,
-            messageExpirationTimeoutGrowFactor: 1.1
-        });
+        ...tests.config,
+        messageExpirationTimeout: 5000,
+        messageExpirationTimeoutGrowFactor: 1.1,
+    });
     let completed = 0;
     const run = async () => {
         const result = await client.contracts.run({